Change: Moved new endpoint to be at /submission/schema/template/:schemaName Code Cleanup: Removed unnecessary code as per PR comments. Code Cleanup: Moved code from template-api into schema-api as per PR comments; deleted template API. Documentation: Updated swagger documentation Change: New endpoint now only returns header rows as per conversation with Rosi. Change: New endpoint now returns Content-type: text/tab-separated-values as per conversation with Jon.
